% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMFDoublet.R
\name{findMfDoublet}
\alias{findMfDoublet}
\title{Find male-female doublet cells}
\usage{
findMfDoublet(x, genome = NULL, qc = FALSE)
}
\arguments{
\item{x}{counts matrix, rows correspond to genes and columns correspond to
cells. Row names must be gene symbols.}

\item{genome}{the genome the data arises from. Current options are
human: genome = "Hs" or mouse: genome = "Mm".}

\item{qc}{logical, indicates whether to perform quality control or not.
qc = TRUE will predict cells that pass quality control only and the filtered
cells will not be classified. qc = FALSE will predict every cell except the
cells with zero counts on *XIST/Xist* and the sum of the Y genes.
Default is TRUE.}
}
\value{
a dataframe with predicted labels for each cell
}
\description{
This function will identify male-female doublet cells in scRNA-seq data. The
classifier is based on random forest models that have been trained
on mouse and human single cell RNA-seq data.
}
\examples{

library(speckle)
library(SingleCellExperiment)
library(CellBench)
library(org.Hs.eg.db)

sc_data <- load_sc_data()
sc_10x <- sc_data$sc_10x

counts <- counts(sc_10x)
ann <- select(org.Hs.eg.db, keys=rownames(sc_10x),
             columns=c("ENSEMBL","SYMBOL"), keytype="ENSEMBL")
m <- match(rownames(counts), ann$ENSEMBL)
rownames(counts) <- ann$SYMBOL[m]

sex <- findMfDoublet(counts, genome="Hs")

table(sex$prediction)
boxplot(counts["XIST",]~sex$prediction)

}
\author{
Xinyi Jin
}
