% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormCounts.R
\name{lognormCounts}
\alias{lognormCounts}
\title{Normalise a counts matrix to the median library size}
\usage{
lognormCounts(x, log = TRUE, prior.count = 0.5, lib.size)
}
\arguments{
\item{x}{a \code{DGEList} object or matrix of counts.}

\item{log}{logical, indicates whether the output should be on the log2 scale
or counts scale. Default is FALSE.}

\item{prior.count}{The prior count to add if the data is log2 normalised.
Default is a small count of 0.5.}

\item{lib.size}{a vector of library sizes to be used during the normalisation
step. Default is NULL and will be computed from the counts matrix.}
}
\value{
a matrix of normalised counts
}
\description{
Normalise a counts matrix to the median library size
}
\author{
Belinda Phipson
}
