% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessDb.R
\name{preprocessDb}
\alias{preprocessDb}
\title{Pre-processing function for Male/Female doublet detection}
\usage{
preprocessDb(x, genome = genome, qc = qc)
}
\arguments{
\item{x}{the counts matrix, rows are genes and columns are cells. Row names
must be gene symbols and column names must be unique.}

\item{genome}{the genome the data arises from. Current options are
human: genome = "Hs" or mouse: genome = "Mm".}

\item{qc}{logical, indicates whether to perform additional quality control
on the cells. qc = TRUE will predict cells that pass quality control only
and the filtered cells will not be classified. qc = FALSE will predict
every cell except the cells with zero counts on *XIST/Xist* and the sum
of the Y genes. Default is TRUE.}
}
\value{
outputs a list object with the following components
\item{tcm.final }{A transposed count matrix where rows are cells and columns
are the features used for classification.}
\item{data.df }{The normalised and scaled \code{tcm.final} matrix.}
\item{discarded.cells }{Character vector of cell IDs for the cells that are
discarded when \code{qc=TRUE}.}
\item{zero.cells }{Character vector of cell IDs for the cells that can not
be classified as male/female due to zero counts on *Xist* and all the
Y chromosome genes.}
}
\description{
The purpose of this function is to process a single cell counts matrix into
the appropriate format for the \code{findMFDoublet} function.
}
\details{
This function will filter out cells that are unable to be classified due to
zero counts on *XIST/Xist* and all of the Y chromosome genes. If
\code{qc=TRUE} additional cells are removed as identified by the
\code{perCellQCMetrics} and \code{quickPerCellQC} functions from the
\code{scuttle} package. The resulting counts matrix is then log-normalised
and scaled.
}
